Also accept dictionaries as ‘query=’ arguments
The URL constructor and the .replace() method now sensibly handle
dictionaries passed to the ‘query’ argument, instead of strictly
requiring it to be an iterable of (key, value) pairs (both of which
must be strings).

Examples:

  URL(query={"foo": "bar"})
  url.replace(query={"foo": "bar"})

This simplifies the API for the common use case, in which:

- the order of the items does not matter
  (though Python >= 3.6 preserves dictionary order)
- no duplicate parameters are present

Updated the docs to include an example of this.

Fixes #50.