Update all dependencies.  Add pypy3 to the matrix.
Note current versions of pytest do not support Python 2.6, 2.7 or 3.4.  That means we need to use pytest 4.6 for those, which means tox needs to be able to select which to use, which means we need to get rid of the static requirements-test.txt, which is probably a good idea anyway.